Refactor `startTestServer` and the systems from testHelpers.go
